added menus adding/removing products functionality added normalize.css organized some tags
