+++ src/main/resources/templates/fragments.html
@@ -4,12 +4,14 @@
       lang="en">
 <head th:fragment="head">
     <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1">
     <title>[[${pageTitle}]] - FastFoodApp</title>
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/css/bootstrap.min.css"
           rel="stylesheet"
           integrity="sha384-EVSTQN3/azprG1Anm3QDgpJLIm9Nao0Yz1ztcQTwFspd3yD65VohhpuuCOmLASjC"
           crossorigin="anonymous">
     <link rel="stylesheet" th:href="@{/styles.css}">
+    <link rel="stylesheet" th:href="@{/normalize.css}">
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/js/bootstrap.bundle.min.js"
             integrity="sha384-MrcW6ZMFYlzcLA8Nl+NtUVF0sA7MsXsP1UyJoMp4YLEuNSfAP+JcXn/tWtIaxVXM"
             crossorigin="anonymous"></script>
@@ -32,9 +34,6 @@
                 </li>
                 <li class="nav-item">
                     <a class="nav-link" th:href="@{/branch/list}">Branches</a>
-                </li>
-                <li class="nav-item">
-                    <a class="nav-link" th:href="@{/}">Order states</a>
                 </li>
             </ul>
         </div>
